feat: Correct the type of parameter and return type

+++ rust-cross-ios/src/rust-cross-ios.h
@@ -1,4 +1,4 @@
 #include <stdint.h>
 
-int add(int a, int b);
-int sub(int a, int b);
+int64_t add(int64_t a, int64_t b);
+int64_t sub(int64_t a, int64_t b);
